bisect(6413): revert justfile+examples from #5389 (compose preserved)
DO NOT MERGE — forensic bisect PR for modular/modular#6413.

Hypothesis: PR #5389 ("make just build actually compile everything")
rewrote the justfile build/test recipes (311-line delta), changed
several example/benchmark Mojo files, and bumped the workflow test
timeout 10 -> 30 min. One of those non-compose changes may have
silenced the libKGEN JIT crash by altering how mojo is invoked,
which test files compile under STRICT, or build ordering.

This PR reverts all #5389 deltas EXCEPT docker-compose.yml. Compose
stays at the green HEAD memlimits (14g/24g/24g) — that variable is
isolated in the companion PR (bisect/6413-revert-compose-memlimit).

Files reverted to pre-#5389 (parent adb4c35c):
  - justfile (recipe rewrite: 278-line delta)
  - .github/workflows/comprehensive-tests.yml (10-line: timeout +
    related)
  - benchmarks/* (4 files, 43-line delta total)
  - examples/* (3 files, minor)
  - docs/dev/mojo-jit-crash-capture-core.md (doc)
  - tests/shared/test_imports.mojo (7-line)

Files preserved at green HEAD:
  - docker-compose.yml (memlimits kept — PR3's variable)
  - All gdb-wrapper plumbing (MOJO_TEST_UNDER_GDB=1 at all 6 sites,
    scripts/mojo-under-gdb.sh, pipe handler)
  - All post-#5389 fixes (e.g. e3e0de83, 9d452948, 5eaf8ad9)

Companion PR (parallel): bisect/6413-revert-compose-memlimit — tests
the +2g dev memlimit alone.

Expected if a non-compose #5389 change was the fix: ~87% failure rate
on Data Utilities + Configs across 8 reruns.
Expected if the rest of #5389 is irrelevant: still green → fix is in
compose memlimit (test via companion PR3) or elsewhere.

+++ benchmarks/scripts/compare_results.mojo
@@ -18,7 +18,7 @@ Output:
 """
 
 from std.sys import argv
-from std.python import Python
+from python import Python
 
 
 # ============================================================================
@@ -123,43 +123,37 @@ def atof(s: String) -> Float64:
     var bytes = s.as_bytes()
 
     # Handle negative numbers
-    if len(bytes) > 0 and bytes[0] == UInt8(ord("-")):
+    if len(bytes) > 0 and bytes[0] == ord("-"):
         sign = -1.0
         i = 1
 
     # Parse integer part
     while (
         i < len(bytes)
-        and bytes[i] != UInt8(ord("."))
-        and bytes[i] != UInt8(ord("e"))
-        and bytes[i] != UInt8(ord("E"))
+        and bytes[i] != ord(".")
+        and bytes[i] != ord("e")
+        and bytes[i] != ord("E")
     ):
         result = result * 10.0 + Float64(Int(bytes[i]) - ord("0"))
         i += 1
 
     # Parse decimal part
-    if i < len(bytes) and bytes[i] == UInt8(ord(".")):
+    if i < len(bytes) and bytes[i] == ord("."):
         i += 1
         var decimal_place: Float64 = 0.1
-        while (
-            i < len(bytes)
-            and bytes[i] != UInt8(ord("e"))
-            and bytes[i] != UInt8(ord("E"))
-        ):
+        while i < len(bytes) and bytes[i] != ord("e") and bytes[i] != ord("E"):
             result = result + Float64(Int(bytes[i]) - ord("0")) * decimal_place
             decimal_place = decimal_place * 0.1
             i += 1
 
     # Parse exponent (scientific notation)
-    if i < len(bytes) and (
-        bytes[i] == UInt8(ord("e")) or bytes[i] == UInt8(ord("E"))
-    ):
+    if i < len(bytes) and (bytes[i] == ord("e") or bytes[i] == ord("E")):
         i += 1
         var exp_sign = 1
-        if i < len(bytes) and bytes[i] == UInt8(ord("-")):
+        if i < len(bytes) and bytes[i] == ord("-"):
             exp_sign = -1
             i += 1
-        elif i < len(bytes) and bytes[i] == UInt8(ord("+")):
+        elif i < len(bytes) and bytes[i] == ord("+"):
             i += 1
 
         var exponent = 0
@@ -333,13 +327,13 @@ def atoi(s: String) -> Int:
     var s_bytes = s.as_bytes()
 
     # Handle negative numbers
-    if len(s_bytes) > 0 and s_bytes[0] == UInt8(ord("-")):
+    if len(s_bytes) > 0 and s_bytes[0] == ord("-"):
         sign = -1
         i = 1
 
     # Parse digits
     while i < len(s_bytes):
-        if s_bytes[i] >= UInt8(ord("0")) and s_bytes[i] <= UInt8(ord("9")):
+        if s_bytes[i] >= ord("0") and s_bytes[i] <= ord("9"):
             result = result * 10 + (Int(s_bytes[i]) - ord("0"))
         i += 1
